backend/requirements.txt
pip
setuptools
wheel

Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
gunicorn==21.2.0
eventlet==0.33.3
python-dotenv==1.0.0

psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

pyaudio==0.2.13
librosa==0.10.0
webrtcvad==2.0.10
azure-cognitiveservices-speech==1.34.0
pydub==0.25.1

torch==2.1.2
transformers==4.36.0
spacy==3.7.2
datasets==2.14.6
peft==0.6.0
accelerate==0.24.1
vllm==0.2.7
scikit-learn==1.3.0
numpy==1.21.6
pandas==1.5.3
pydantic==1.10.13

TTS==0.13.1

kafka-python==2.0.2
confluent-kafka==2.3.0

prometheus-client==0.19.0
structlog==23.2.0

requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
